feat(server): :sparkles: Add Server-Side Decorations
In hindsight the extra scene tree for the layer we place toplevels on in a workspace was really stupid, and made it really hard to implement this, so it was redone from scratch. this also allowed for some nice optimization in the worksapce logic. We put the toplevel and the decoration into a container node, which we can then move easily.