fix(server): :bug: Fix segfault on closing of compositor
i forgot i hacked together the closing procedure, which used to init a new server var so when you try to close it it gets all fucked uo and segfaults. there was also a problem where you could segfault in some very specific cases when you close a window while dragging it because i forgot an extra check.